// src/anotacoes/dto/anotacoes.dto.ts
export interface AnotacoesDto {
  id: number;
  usuario: number;
  nota1: string;
  nota2: string;
  nota3: string;
  nota4: string;
  nota5: string;
  nota6: string;
  nota7: string;
  nota8: string;
  nota9: string;
  nota10: string;
  tituloNota1: string;
  tituloNota2: string;
  tituloNota3: string;
  tituloNota4: string;
  tituloNota5: string;
  tituloNota6: string;
  tituloNota7: string;
  tituloNota8: string;
  tituloNota9: string;
  tituloNota10: string;
  dataNota1: string;
  dataNota2: string;
  dataNota3: string;
  dataNota4: string;
  dataNota5: string;
  dataNota6: string;
  dataNota7: string;
  dataNota8: string;
  dataNota9: string;
  dataNota10: string;
}